requirements.txt
langchain
langgraph
fastapi
fastapi-cli
python-dotenv
fire
unstructured
langchain_community
lancedb
faiss-cpu
pyarrow
tiktoken
openai
langchainhub
langchain-openai
PyPDF2
PyYAML
arxiv
streamlit